+++ dbms/src/Storages/Page/universal/UniversalPageStorage.cpp
@@ -12,12 +12,12 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include <Storages/Page/UniversalPageStorage.h>
 #include <Storages/Page/V3/Blob/BlobConfig.h>
 #include <Storages/Page/V3/BlobStore.h>
 #include <Storages/Page/V3/PageDirectoryFactory.h>
 #include <Storages/Page/V3/PageStorageImpl.h>
 #include <Storages/Page/V3/WAL/WALConfig.h>
+#include <Storages/Page/universal/UniversalPageStorage.h>
 
 namespace DB
 {
@@ -25,16 +25,15 @@ namespace DB
 UniversalPageStoragePtr UniversalPageStorage::create(
     String name,
     PSDiskDelegatorPtr delegator,
-    const UniversalPageStorage::Config & config,
+    const PageStorageConfig & config,
     const FileProviderPtr & file_provider)
 {
     UniversalPageStoragePtr storage = std::make_shared<UniversalPageStorage>(name, delegator, config, file_provider);
-    PS::V3::BlobConfig blob_config; // FIXME: parse from config
     storage->blob_store = std::make_shared<PS::V3::BlobStore<PS::V3::universal::BlobStoreTrait>>(
         name,
         file_provider,
         delegator,
-        blob_config);
+        PS::V3::BlobConfig::from(config));
     return storage;
 }
 
@@ -43,8 +42,9 @@ void UniversalPageStorage::restore()
     blob_store->registerPaths();
 
     PS::V3::universal::PageDirectoryFactory factory;
-    PS::V3::WALConfig wal_config; // FIXME: parse from config
-    page_directory = factory.setBlobStore(*blob_store).create(storage_name, file_provider, delegator, wal_config);
+    page_directory = factory
+                         .setBlobStore(*blob_store)
+                         .create(storage_name, file_provider, delegator, PS::V3::WALConfig::from(config));
 }
 
 void UniversalPageStorage::write(UniversalWriteBatch && write_batch, const WriteLimiterPtr & write_limiter) const
